+++ spec/units/evil_events/core/events/serializers/json_spec.rb
@@ -1,22 +1,10 @@
 # frozen_string_literal: true
 
 describe EvilEvents::Core::Events::Serializers::JSON, :stub_event_system do
-  it_behaves_like 'event serializer component' do
-    let(:serializer) { EvilEvents::Core::Events::Serializers::JSON::Factory.new.create! }
-    let(:serialization_error) { EvilEvents::JSONSerializationError }
-    let(:deserialization_error) { EvilEvents::JSONDeserializationError }
-    let(:serialization_type) { String }
-    let(:incorrect_deserialization_objects) { gen_all }
-    let(:invalid_serializations) do
-      data = serializer.serialize(event)
+  include_context 'event system'
 
-      [
-        data.sub('"type":', "\"#{gen_str}\":"),
-        data.sub('"payload":', "\"#{gen_str}\":"),
-        data.sub('"metadata":', "\"#{gen_str}\":")
-      ]
-    end
-
-    specify { expect(serializer).to be_a(described_class) }
+  context 'native engine' do
+    before { system_config.configure { |c| c.serializers.json.engine = :native } }
+    it_behaves_like 'json event serialization component'
   end
 end